# erlang-server/generated/priv/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Jenkins API clients generated from Swagger / Open API specification",
    "version" : "0.1.0",
    "title" : "Swaggy Jenkins",
    "contact" : {
      "name" : "Cliffano Subagio",
      "url" : "http://github.com/cliffano/swaggy-jenkins",
      "email" : "blah@cliffano.com"
    }
  },
  "basePath" : "/",
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/blue/rest/classes/" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Get classes details",
        "operationId" : "search",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "description" : "Query string containing an array of class names",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved search result",
            "schema" : {
              "type" : "string"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/classes/{class}" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Get a list of class names supported by a given class",
        "operationId" : "getClasses",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "class",
          "in" : "path",
          "description" : "Name of the class",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved class names",
            "schema" : {
              "type" : "string"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/search/" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Search for any resource details",
        "operationId" : "search",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "description" : "Query string",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved search result",
            "schema" : {
              "type" : "string"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/organizations/" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve all organisations details",
        "operationId" : "getOrganisations",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved pipelines details",
            "schema" : {
              "$ref" : "#/definitions/getOrganisations"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve organisation details",
        "operationId" : "getOrganisation",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved pipeline details",
            "schema" : {
              "$ref" : "#/definitions/swaggyjenkinsOrganisation"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          },
          "404" : {
            "description" : "Pipeline cannot be found on Jenkins instance"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}/pipelines/" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve all pipelines details for an organisation",
        "operationId" : "getPipelinesByOrg",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved pipelines details",
            "schema" : {
              "$ref" : "#/definitions/getPipelines"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}/pipelines/{pipeline}" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve pipeline details for an organisation",
        "operationId" : "getPipelineByOrg",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pipeline",
          "in" : "path",
          "description" : "Name of the pipeline",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved pipeline details",
            "schema" : {
              "$ref" : "#/definitions/swaggyjenkinsPipeline"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          },
          "404" : {
            "description" : "Pipeline cannot be found on Jenkins instance"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}/pipelines/{pipeline}/branches" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve all branches details for an organisation pipeline",
        "operationId" : "getPipelineBranchesByOrg",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pipeline",
          "in" : "path",
          "description" : "Name of the pipeline",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved all branches details",
            "schema" : {
              "$ref" : "#/definitions/getMultibranchPipeline"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}/pipelines/{pipeline}/branches/{branch}/" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve branch details for an organisation pipeline",
        "operationId" : "getPipelineBranchByOrg",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pipeline",
          "in" : "path",
          "description" : "Name of the pipeline",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "branch",
          "in" : "path",
          "description" : "Name of the branch",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved branch details",
            "schema" : {
              "$ref" : "#/definitions/iojenkinsblueoceanrestimplpipelineBranchImpl"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}/pipelines/{folder}/" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve pipeline folder for an organisation",
        "operationId" : "getPipelineFolderByOrg",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folder",
          "in" : "path",
          "description" : "Name of the folder",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved folder details",
            "schema" : {
              "$ref" : "#/definitions/iojenkinsblueoceanserviceembeddedrestPipelineFolderImpl"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}/pipelines/{folder}/pipelines/{pipeline}" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve pipeline details for an organisation folder",
        "operationId" : "getPipelineFolderByOrg",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "pipeline",
          "in" : "path",
          "description" : "Name of the pipeline",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folder",
          "in" : "path",
          "description" : "Name of the folder",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved pipeline details",
            "schema" : {
              "$ref" : "#/definitions/iojenkinsblueoceanserviceembeddedrestPipelineImpl"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}/user/" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve authenticated user details for an organisation",
        "operationId" : "getAuthenticatedUser",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved authenticated user details",
            "schema" : {
              "$ref" : "#/definitions/swaggyjenkinsUser"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}/users/" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve users details for an organisation",
        "operationId" : "getUsers",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved users details",
            "schema" : {
              "$ref" : "#/definitions/swaggyjenkinsUser"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/blue/rest/organizations/{organisation}/users/{user}" : {
      "get" : {
        "tags" : [ "blueOcean" ],
        "description" : "Retrieve user details for an organisation",
        "operationId" : "getUser",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "organisation",
          "in" : "path",
          "description" : "Name of the organisation",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "user",
          "in" : "path",
          "description" : "Name of the user",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved users details",
            "schema" : {
              "$ref" : "#/definitions/swaggyjenkinsUser"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "jenkins_auth" : {
      "type" : "basic"
    }
  },
  "definitions" : {
...
    "getMultibranchPipeline" : {
      "type" : "object",
      "properties" : {
        "displayName" : {
          "type" : "string"
        },
        "estimatedDurationInMillis" : {
          "type" : "integer"
        },
        "latestRun" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "organization" : {
          "type" : "string"
        },
        "weatherScore" : {
          "type" : "integer"
        },
        "branchNames" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "numberOfFailingBranches" : {
          "type" : "integer"
        },
        "numberOfFailingPullRequests" : {
          "type" : "integer"
        },
        "numberOfSuccessfulBranches" : {
          "type" : "integer"
        },
        "numberOfSuccessfulPullRequests" : {
          "type" : "integer"
        },
        "totalNumberOfBranches" : {
          "type" : "integer"
        },
        "totalNumberOfPullRequests" : {
          "type" : "integer"
        },
        "_class" : {
          "type" : "string"
        }
      }
    },
...
    "swaggyjenkinsOrganisation" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
    "getOrganisations" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/swaggyjenkinsOrganisation"
      }
    },
...
    "swaggyjenkinsPipeline" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "organization" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "fullName" : {
          "type" : "string"
        },
        "weatherScore" : {
          "type" : "integer"
        },
        "estimatedDurationInMillis" : {
          "type" : "integer"
        },
        "latestRun" : {
          "$ref" : "#/definitions/swaggyjenkinsPipeline_latestRun"
        }
      }
    },
...
    "iojenkinsblueoceanrestimplpipelineBranchImpl" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "estimatedDurationInMillis" : {
          "type" : "integer"
        },
        "fullDisplayName" : {
          "type" : "string"
        },
        "fullName" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "organization" : {
          "type" : "string"
        },
        "parameters" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelStringParameterDefinition"
          }
        },
        "permissions" : {
          "$ref" : "#/definitions/iojenkinsblueoceanrestimplpipelineBranchImpl_permissions"
        },
        "weatherScore" : {
          "type" : "integer"
        },
        "pullRequest" : {
          "type" : "string"
        }
      }
    },
...
    "getPipelines" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/swaggyjenkinsPipeline"
      }
    },
    "swaggyjenkinsUser" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "fullName" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
...
  },
...
}